clean constnats import
