+++ packages/lib/index.js
@@ -25,5 +25,6 @@ export * from "./src/dto";
 export * from "./src/plandetransport";
 export * from "./src/cohort";
 export * from "./src/constants/cle/classeConstants";
+export * from "./src/constants/email";
 
 export * from "./src/constants/referentConstants";
