+++ packages/lib/src/constants/email.ts
@@ -1,0 +1,204 @@
+import { ROLES, SUB_ROLES } from "../roles";
+
+const SENDINBLUE_TEMPLATES = {
+  // Auth
+  SIGNIN_2FA: "1317",
+  SIGNUP_EMAIL_VALIDATION: "1316",
+  PROFILE_EMAIL_VALIDATION: "1347",
+  FORGOT_PASSWORD: "1308",
+  INVITATION_CHEF_ETABLISSEMENT_TO_INSCRIPTION_TEMPLATE: "1977",
+  INVITATION_CHEF_ETABLISSEMENT_TO_CONFIRMATION_TEMPLATE: "1978",
+  invitationReferent: {
+    [ROLES.REFERENT_DEPARTMENT]: "158",
+    [ROLES.REFERENT_REGION]: "159",
+    [ROLES.RESPONSIBLE]: "160",
+    [ROLES.SUPERVISOR]: "160",
+    [ROLES.ADMIN]: "161",
+    [ROLES.HEAD_CENTER]: "162",
+    [ROLES.VISITOR]: "286",
+    [ROLES.DSNJ]: "662",
+    [ROLES.TRANSPORTER]: "662",
+    [ROLES.ADMINISTRATEUR_CLE]: "1977",
+    [ROLES.REFERENT_CLASSE]: "1391",
+    [SUB_ROLES.coordinateur_cle]: "1390",
+    NEW_STRUCTURE: "160",
+    NEW_STRUCTURE_MEMBER: "163",
+  },
+  INVITATION_YOUNG: "1283",
+
+  //session phase 1
+  SHARE_SESSION_PHASE1: "419",
+
+  //PHASE 2
+  ATTACHEMENT_PHASE_2_APPLICATION: "1299",
+
+  // contract
+  VALIDATE_CONTRACT: "1288",
+  REVALIDATE_CONTRACT: "1287",
+
+  referent: {
+    WELCOME_REF_DEP: "378",
+    WELCOME_REF_REG: "391",
+    YOUNG_CHANGE_COHORT: "324",
+    YOUNG_CHANGE_COHORT_HTS_TO_CLE: "1462",
+    YOUNG_CHANGE_COHORT_CLE_TO_HTS: "1463",
+    DEPARTURE_CENTER: "935",
+
+    RECAP_BI_HEBDO_DEPARTMENT: "231",
+    // MIG
+    MISSION_WAITING_CORRECTION: "164",
+    MISSION_WAITING_VALIDATION: "194",
+    MISSION_VALIDATED: "63",
+    MISSION_END: "213",
+    MISSION_CANCEL: "233",
+    NEW_MISSION: "192",
+    NEW_MISSION_REMINDER: "195",
+    MISSION_REFUSED: "165",
+    CANCEL_APPLICATION: "155",
+    CANCEL_APPLICATION_PHASE_2_VALIDATED: "599",
+    ABANDON_APPLICATION: "214",
+    VALIDATE_APPLICATION_TUTOR: "196",
+    NEW_APPLICATION_MIG: "156",
+    YOUNG_VALIDATED: "173",
+    APPLICATION_REMINDER: "197",
+
+    CONTRACT_DRAFT: "199",
+
+    STRUCTURE_REGISTERED: "191",
+
+    MISSION_ARCHIVED: "204",
+    MISSION_ARCHIVED_1_WEEK_NOTICE: "205",
+
+    // PREPA MILITAIRE
+    MILITARY_PREPARATION_DOCS_SUBMITTED: "149",
+    MILITARY_PREPARATION_DOCS_VALIDATED: "148",
+
+    NEW_APPLICATION: "new-application",
+    RELANCE_APPLICATION: "relance-application",
+    //PHASE 3
+    VALIDATE_MISSION_PHASE3: "174",
+
+    // Support
+    MESSAGE_NOTIFICATION: "218",
+
+    //EQUIVALENCE
+    EQUIVALENCE_WAITING_VERIFICATION: "341",
+
+    // Account deletion
+    DELETE_ACCOUNT_NOTIFICATION_1: "615",
+    DELETE_ACCOUNT_NOTIFICATION_2: "616",
+    YOUNG_WITHDRAWN_CLE: "1741",
+
+    INSTRUCTION_END_REMINDER: 874,
+  },
+  young: {
+    CHANGE_COHORT: "1461",
+    WAITING_LIST: "1215",
+    WITHDRAWN: "1248",
+    INSCRIPTION_REMINDER: "1276",
+    INSCRIPTION_STARTED: "1277",
+    INSCRIPTION_VALIDATED: "1284",
+    INSCRIPTION_VALIDATED_CLE: "1399",
+    INSCRIPTION_REACTIVATED: "1247",
+    INSCRIPTION_WAITING_CORRECTION: "1285",
+    INSCRIPTION_REMIND_CORRECTION: "1285",
+    INSCRIPTION_WAITING_LIST: "1286",
+    INSCRIPTION_REFUSED: "1246",
+    INSCRIPTION_END_IN_PROGRESS_ELIGIBLE: "1271",
+    INSCRIPTION_END_IN_PROGRESS_NOT_ELIGIBLE: "1270",
+    INSTRUCTION_END_WAITING_VALIDATION_ELIGIBLE: "978",
+    INSTRUCTION_END_WAITING_VALIDATION_NOT_ELIGIBLE: "1268",
+    INSTRUCTION_END_WAITING_CORRECTION_ELIGIBLE: "936",
+    INSTRUCTION_END_WAITING_CORRECTION_NOT_ELIGIBLE: "1267",
+    PARENT_CONSENTED: "1303",
+    PARENT_DID_NOT_CONSENT: "1304",
+    PARENT2_DID_NOT_CONSENT: "1243",
+    PHASE1_AFFECTATION: "1272",
+    INSCRIPTION_WAITING_CONSENT: "1245",
+
+    // MIG
+    REFUSE_APPLICATION: "1253",
+    CANCEL_APPLICATION: "1262",
+    VALIDATE_APPLICATION: "1257",
+    MISSION_PROPOSITION: "1229",
+    MISSION_CANCEL: "1236",
+    MISSION_ARCHIVED: "1274",
+    MISSION_ARCHIVED_AUTO: "1289",
+    PHASE_2_VALIDATED: "1256",
+    MISSION_PROPOSITION_AUTO: "1234",
+
+    // PREPA MILITAIRE
+    MILITARY_PREPARATION_DOCS_VALIDATED: "1258",
+    MILITARY_PREPARATION_DOCS_CORRECTION: "1259",
+    MILITARY_PREPARATION_DOCS_REFUSED: "1260",
+
+    //PHASE 3
+    VALIDATE_PHASE3: "1280",
+
+    DOCUMENT: "1310",
+    CONTRACT_VALIDATED: "1282",
+
+    // Personal and situation changes
+    DEPARTMENT_IN: "215",
+    DEPARTMENT_OUT: "216",
+
+    //Phase 1 pj
+    PHASE_1_PJ_WAITING_VERIFICATION: "1239",
+    PHASE_1_PJ_WAITING_CORRECTION: "1293",
+    PHASE_1_PJ_VALIDATED: "1294",
+    PHASE_1_FOLLOW_UP_MEDICAL_FILE: "1297",
+    PHASE_1_AFFECTATION: "1298",
+    PHASE_1_WAITING_CORRECTION_REMINDER: "1275",
+
+    //send a download link to the young
+    LINK: "1231",
+
+    //EQUIVALENCE
+    EQUIVALENCE_REFUSED: "1249",
+    EQUIVALENCE_VALIDATED: "1251",
+    EQUIVALENCE_WAITING_VERIFICATION: "1292",
+    EQUIVALENCE_WAITING_CORRECTION: "1250",
+
+    APPLICATION_CANCEL_1_WEEK_NOTICE: "1281",
+    APPLICATION_CANCEL_13_DAY_NOTICE: "1230",
+
+    // CLE
+    PHASE_1_ATTESTATION_SWITCH_CLE: "1659",
+  },
+  YOUNG_ARRIVED_IN_CENTER_TO_REPRESENTANT_LEGAL: "1290",
+  parent: {
+    OUTDATED_ID_PROOF: "1302",
+    PARENT1_REVALIDATE_RI: "1579",
+    PARENT1_CONSENT: "1300",
+    PARENT2_CONSENT: "1301",
+    PARENT1_CONSENT_REMINDER: "1305",
+    PARENT_YOUNG_COHORT_CHANGE: "1307",
+    PARENT1_RESEND_IMAGERIGHT: "1241",
+    PARENT2_RESEND_IMAGERIGHT: "1241",
+    PARENT2_IMAGERIGHT_REMINDER: "1309",
+  },
+  headCenter: {
+    FILE_SESSION_REMINDER: "1232",
+  },
+  PLAN_TRANSPORT: {
+    MODIFICATION_REFUSEE: "673",
+    DEMANDE_DE_MODIFICATION: "672",
+    MODIFICATION_ACCEPTEE: "674",
+    MODIFICATION_SCHEMA: "810",
+    NOTIF_REF: "1131",
+    NOTIF_TRANSPORTEUR_MODIF: "1502",
+  },
+  CLE: {
+    CLASSE_CREATED: "1354",
+    CLASSE_INFOS_COMPLETED: "1414",
+    CLASSE_INFOS_COMPLETED_DEP_REG: "1412",
+    CLASSE_VALIDATED: "1429",
+    CONFIRM_SIGNUP_COORDINATEUR: "1413",
+    CONFIRM_SIGNUP_REFERENT_ETABLISSEMENT: "1395",
+    CONFIRM_SIGNUP_REFERENT_CLASSE: "1396",
+    REFERENT_AFFECTED_TO_CLASSE: "1427",
+    CLASSE_COHORT_UPDATED: "1460",
+  },
+};
+
+export { SENDINBLUE_TEMPLATES };
